topic-mention: Modify typeaheads to improve completion experience.
Earlier, we would place the cursor before the ending `**` of the stream
completion. This would mean that backspacing on the completion wouldn't
retrigger the stream typeahead eventually. To resolve, we remove the
trailing `**` on topic_jump and add them back on topic_list selection.

Also, we implement partial completion of stream typeahead on pressing
'>'. We use our custom addition to typeahead.js: this.trigger_selection
to start the topic_list typeahead.

At the end of this commit, we support the following mechanisms to
complete the stream-topic mention:

1. Type "#denmar|".
2. Press Enter to get "#**Denmark** |".
3. Press > to get "#**Denmark>|".
4. Type topic name and press enter.

OR

1. Type "#denmar|".
2. Type > to get "#**Denmark>|".
3. Type topic name and press enter.

Both result in the final inserted sytax: "#**Denmark>topic name**".

Also makes minor changes to tests.